refactor(runtime-tags): merge patch-write channels; side-effect-free admission guard; derived patch decisions
