+++ packages/runtime-tags/src/translator/util/known-tag.ts
@@ -27,7 +27,7 @@ import {
   some,
   toIter,
 } from "./optional";
-import { kInstancePatchSkip } from "./persisted/delivery";
+import { getChildPatchPlan } from "./persisted/decisions";
 import { addPersistedChildRenderer } from "./persisted/intrinsics";
 import { onFinalizePersisted } from "./persisted/lifecycle";
 import {
@@ -250,8 +250,11 @@ export function knownTagTranslateHTML(
   if (isPersisted()) addPersistedChildRenderer(tagIdentifier);
   // A client-owned instance renders nothing into a patch: the link and the
   // child render skip together, and the absent entry keeps the live child.
-  let clientOwnedStatements: t.Statement[] | undefined =
-    isPersisted() && tagExtra[kInstancePatchSkip] ? [] : undefined;
+  const skipsPatchRender =
+    isPersisted() && getChildPatchPlan(tag).skipsPatchRender;
+  let clientOwnedStatements: t.Statement[] | undefined = skipsPatchRender
+    ? []
+    : undefined;
 
   let varStatement: t.Statement | undefined;
   if (childScopeSerializeReason) {
@@ -279,7 +282,7 @@ export function knownTagTranslateHTML(
           peekScopeId,
         ),
       );
-      if (tagExtra[kInstancePatchSkip]) {
+      if (skipsPatchRender) {
         clientOwnedStatements = [patchChildStatement];
       } else {
         statements.push(patchChildStatement);
